+++ docs/ChangeLog.md
@@ -137,6 +137,7 @@
 - edit: support a user command `ble append-line` (requested by mozirilla213) `#D2001` 2a524f34
 - decode: accept isolated <kbd>ESC \<char\></kbd> (requested by mozirilla213) `#D2004` xxxxxxxx
 - sabbrev: add widget `magic-slash` to approximate Zsh named directories (motivated by mozirilla213) `#D2008` xxxxxxxx
+- sabbrev: support inline and linewise sabbre with `ble-sabbrev -il` `#D2012` xxxxxxxx
 
 ## Changes